+++ src/rdkafka_feature.c
@@ -46,9 +46,7 @@ static const char *rd_kafka_feature_names[] = {
         "IdempotentProducer",
         "ZSTD",
         "SaslAuthReq",
-#ifdef RD_KAFKA_FEATURE_KIP360
-        "KIP-360",
-#endif
+        "KIP360",
         "UnitTest",
         NULL
 };
@@ -204,17 +202,15 @@ static const struct rd_kafka_feature_map {
                         { -1 },
                 },
         },
-#ifdef RD_KAFKA_FEATURE_KIP360
         {
                 /* @brief >=2.4.0: KIP-360 */
                 .feature = RD_KAFKA_FEATURE_KIP360,
                 .depends = {
-                        { RD_KAFKAP_InitProducerId, 2, 2 },
+                        { RD_KAFKAP_InitProducerId, 3, 3 },
                         { -1 },
                 },
 
         },
-#endif
         { .feature = 0 }, /* sentinel */
 };